In preparation for introduction of voting FSM, extract class instance
We need a header that provides a public raft API for the clients
to use. The best candidate is raft.hh.

We also should ideally separate Raft protocol finite state
machine from the implementation details of recovery, snapshotting, as well
as keeping peers and followers up to date.

Begin the spit by moving class instance declaration
to its own header.