Implement flow control.
Each follower can be in PROBE or PIPELINE state. In the former case only
one outstanding entry is send each tick, in the later up to 10 entries
are sent (will make it size based later). A follower moves from probe to
pipeline on a match and back on a mismatch. To ensure forward progress
in face of a packet lose on each tick additional packet is sent if
needed otherwise keepalive is sent to ensure that leadership is
maintained.