Harmonize API use of references.
Prefer values to references in public API - this makes
the API robust.
The caller can move its object into the argument
if the caller doesn't need the object. Otherwise we'll take a copy,
which is safe & laconic (while not as fast as rvalue-reference, in
public API the speed difference is not critical).

In internal API, expect rvalue-reference, rather than value.

Note, that it's safe to bind both const and rvalue references
to temporary thanks to lifetime extension feature of C++.