raft: joint consensus, don't store configuration in FSM
In follower state, FSM doesn't know the current cluster
configuration.  Instead of trying to watch the follower log for
configuration changes to keep FSM copy up to date, remove it from
FSM altogether since the follower doesn't need it anyway.

When entering candidate or leader state, fetch the most recent
configuration from the log and initialize the state specific
state with it.