raft: testing: fsm set_max_election_timeout()
Raft nodes set a _randomized_election_timeout per term. It lies in the
range [ELECTION_TIMEOUT + 1, 2 * ELECTION_TIMEOUT].

To give tests with custom elections a wider margin, we provide a method
to bump the election timeout to the maximum value. This is to be called
in the the servers intended to be followers after the election. This way
we assure they won't have a chance to become candidates by a tick.
(i.e. avoiding having them with the minimum of this range
_randomized_election_timeout == ELECTION_TIMEOUT + 1
and by chance getting ticked right away would make them candidates)

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>