commitlog: Make force_sync parameter optionally a timeout and allow periodic wait
Caller can now initialize force_sync with a time_point representing
the latest we accept data to be flushed. Doing so will also wait
for said flush to happen.

If set, and flush task would run later than the time, we change the
timer execution time. This will allow more writes to enter the current
segment, to avoid the space waste of frequent flushing associated with
raw batch mode (or force_sync::yes). If time already passed we do the
sync op directly.