raft: pass snapshot_reply into fsm::step()
By the time we receive snapshot_reply from a follower
we may no longer be a leader. Follower term may be
different from snapshot term, e.g. the follower may
be aware of a new leader already and have a higher term.

We should pass this information into FSM via so that it
can correctly change its state, and not call FSM directly.