raft: only apply configuration once it was persisted
The state machine will send out messages based on the new configuration.
If the leader later restarts and the configuration entry is not
present in the log, this breaks linearizability of the observable
state machine state changes.

Besides, if the configuration doesn't contain the current leader, this
leader is still responsible for carrying out the transition to the new
configuration - i.e.  committing it on the majority of the nodes,
and then converting to a follower to not disturb the new cluster.
Something it can't do reliably without first persisting it locally.

Clear the current leader on the follower if it is removed from the
configuration. In vanilla raft this would be taken care of the failure
detection based on pings - the leader would stop pinging its
followers after removing itself from the configuration. Since
Seastar Raft relies on a shared failure detector, we must keep
track of the configuration in the log instead.

The issues are found when adding {A} -> {B} configuration change
transition test, which comes as a follow up.

Improve configuration changes tracing.