raft: replication test: support config change
Add support for configuration change on leader.

Keep track of servers in config in test.

Add a dummy entry to confirm configuration changed. If the add fails,
because the old leader was not in the new config and stepped down, the
config is considered changed, too.

Add a test cycling every scenario for one of the 4 nodes.
NOTE: this test keeps quorum + 1 as previous node need to be
disconnected.

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>