raft: replication test: fix custom election
Use the new specific connectivity to manage old leader disconnection
more specifically.

This fixes having elections where the vote of the old leader is required
for quorum. For example {A,B} and we want to switch leader. (@gleb-cloudius)

The existing etcd test workaround (n=3) is corrected back to original
(n=2).

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>