raft: (server) implement taking initial snapshot at start
Raft instance should never run without a snapshot and configuration,
so implement a way to take a snapshot for new servers before
starting communication.

This is also necessary to make creation of a Raft instance idempotent
during Scylla cluster discovery. Imagine a fresh cluster setup of two
nodes {A,B}, with a single seed pointing to node A.
Node A boots, creates a Raft group id and initial Raft configuration
with {A} in it, but fails to persist the snapshot. Node B learns group
id from A and persists it locally.
Then A restarts, discovers no persistent state for its Group 0, creates
a new group 0 id and a new group 0 snapshot.