raft: replication test: fix elect_new_leader old leader connection
Recently, the logic of elect_new_leader was changed to allow the old
leader to vote for the new candidate. But the implementation is wrong as
it re-connects the old leader in all cases disregarding if the nodes
were already disconnected.

Check if both old leader and the requested new leader are connected
first and only if it is the case then the old leader can participate in
the election.

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>