raft: give precandidate at least 2 ticks to run election
Servers have election timeout (10) + random ticks (1-10) to become
candidates if no leader is seen.

When there are many servers (e.g. hundreds) candidates can show up
constantly and prevent an election to happen.

This patch gives 2 extra ticks of space to a valid prevote candidate.

The worst trade-off is a couple of ticks more in case of failed election
for whatever reason, in the case of very few servers (e.g. 10). For
higher numbers (e.g. hundreds) the difference is negligible.

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>