raft: give precandidate at least one tick to run election
Servers have election timeout (10) + random ticks (1-10) to become
candidates if no leader is seen.

When there are hundreds of servers there is a high chance of another
server becoming a candidate while there is already a valid candidate
trying to get elected. At certain point this becomes too disruptive and
elections never resolve.

This patch gives a valid precandidate at least a whole tick time to run
the election.

This follower backoff is enabled for configurations of 100+ servers.

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>