raft: election timeout with exponential distribution
When a leader drops the nodes wait election timeout (10 ticks) to be
receptive to a candidate. And an extra amount of ticks to become
candidates themselves. This extra is set with a uniform distribution.

For configurations of hundreds of nodes this implies at the 11th tick
there will be about ~10% of candidates simultaneously. But since the
ticks are not precisely aligned, within the 11th tick there is some
time. But as there are more and more nodes even the tick time becomes
crowded and elections can become hung forever as more and more nodes
become candidates.

This patch proposes a non-uniform distribution of candidate timeouts,
making smaller values less likely. This way there is a higher chance of
having a successful election in the first two ticks.

The trade-off is for very small configurations (i.e. <10 nodes) the
average election timeout grows by ~5 ticks. But this difference becomes
negligible at larger configurations (i.e. >50 nodes).

From experiments, with a lambda of 0.3, this change allows clusters of
up to 500 nodes to avoid hung elections.

Combined with prevote follower backoff of 2 ticks, clusters of up to 750
nodes behave properly.

A higher lambda might help with larger clusters (i.e. 1000 nodes) were
there are still problems.

https://gist.github.com/alecco/7ed7808095afd4f98b90f808e29723cc

lambda == 0.3
 01:  1% 00173 ***
 02:  2% 00242 ****
 03:  3% 00311 ******
 04:  4% 00437 ********
 05:  6% 00619 ************
 06:  8% 00817 ****************
 07: 11% 01165 ***********************
 08: 14% 01472 *****************************
 09: 20% 02020 ****************************************
 10: 27% 02744 ******************************************************
useful samples 70%

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>