raft: candidate timeout proportional to cluster size
To avoid dueling candidates with large clusters, make the timeout
proportional to the cluster size.

Debug mode is too slow for a test of 1000 nodes so it's disabled, but
the test passes for release and dev modes.

For release mode use 200ms to allow messages to be processed on time.

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>