raft: (server) allow adding entries/modify config on a follower
Implement an RPC to forward add_entry calls from the follower
to leader. Bounce & retry in case of not_a_leader.
Do not retry in case of uncertainty - this can lead to adding
duplicate entreis.

When forwarding an entry to a remote leader we may get back
a term/index pair that conflicts (has the same index, but is with
a higher term) with a local entry we're still waiting on.

This can happen, e.g. because there was a leader change and the
log was truncated, but we still haven't got the append_entries
RPC from the new leader, still haven't truncated the log locally, still
haven't aborted all the local waits for truncated entries.

Only remove the offending entry from the wait list and abort it. There
may be entries labeled with an older term on the left (with a lower commit index)
as well as on the right (with a higher commit entry) of the conflicting
entry. However, finding them, in either direction, would require a
linear scan. If we allow it, we may end up doing this linear scan for
*every* conflicting entry during the transition period, which already
brings us to N^2 complexity of this step. At the same time, as soon as
append_entries RPC from the leader reaches this follower the waits
for the truncated part of the log will be aborted anyway, so the scan
is unnecessary.

Similarly to being able to add entries, allow to modify
Raft group configuration on a follower. The implementation
works the same way as adding entries - forwards the command
to the leader.

Now that add_entry() may time out handle timed_out_error
explicitly on RPC level.