raft: (server) implement add/remove server RPC
Add/remove server is used in Raft topology changes. Since
server ids are unique, it's OK to retry these operations.
The implementation provides an easy to use wrapper
around set_configuration() to add/remove
servers on a follower by forwarding the request to the leader,
retry in case of error, wait for leader if it's unknown.