raft: timeuuid in system_schema.scylla_tables
WIP schema change with canonical mutation to raft log
Add 2 columns to keep timeuuid of schema changes.
This is needed to match serialized change mutation in Raft log.
The list of previous timeuuid is a frozen mutation as it needs to be
checked before updating it.
raft: WIP store mutation of schema change
Generate timeuuid, create mutation, store it.