test: raft: randomized_nemesis_test: persistence_wrapper: store a shared pointer to `persistence`
We want the test to be able to reuse `persistence` even after
`persistence_wrapper` is destroyed for simulating server restarts. We'll
do it by having the test keep a shared pointer to `persistence`.

To do that, instead of storing `persistence` by value and constructing
it inside `persistence_wrapper`, store it by `lw_shared_ptr` which is
taken through the constructor (so `persistence` itself is now
constructed outside of `persistence_wrapper`).