service: raft: schema_raft_state_machine: check and update state ID in `apply`
The schema state machine will only modify state during command
application if the provided "previous schema ID" is equal to the
last schema ID present in the history table. Otherwise, the command will
be a no-op.

To ensure linearizability of schema changes, the performer of the schema
change must first read the last schema ID, only then read the schema state
and send a command for the state machine. If a concurrent schema change
races with this command and manages to modify the state, we will detect
that the last schema ID does not match during `apply`; all calls to
`apply` are serialized, and `apply` adds the new entry to the history
table at the end, after modifying the schema state.

There's still some work to do for this idea to work:
1. We need to ensure that these linearized schema changes don't race
   with calls of `announce_unconditionally`.
2. We need to include the history table in snapshot transfers (following
   commit).