service: migration_manager: introduce `announce_unconditionally`
Regular schema changes, such as performed by the user using a CQL DDL
statement, will work as follows:
1. read the schema state ID
2. validate if the change is possible with the current schema
3. construct the mutations
4. send a command that contains the mutations, the obtained state ID,
   and a new state ID.

When the command is applied, it will apply the mutations only if the
state ID read in step 1 is still equal to the current schema state ID.
If not, it means there was a concurrent schema change applied earlier
which may have invalidated our mutations, so we cannot apply them.

However we may want to leave a "backdoor" which allows applying schema
changes without the state ID validation. We can use this backdoor e.g.
to create internal distributed system tables which by design do not
conflict with each other, and should also not conflict with user issued
DDL statements as they are created when the cluster is first created and
CQL is made available. We can also use this backdoor temporarily for
subsystems other than CQL before they migrate to using the "correct" way
(e.g. alternator, thrift).

This will be the purpose of `announce_unconditionally`. For now it
simply calls `announce`, but this will change in following commits.