test.py: speed up the search for test cases
Case cache was designed to speed up case search for --repeat option,
while it'd be nice to use across different build modes as well.
"debug" is lexicographically smaller than "dev" and "release"
and we always begin populating the case list using debug builds, while
this takes the longest.

In fact, if we scan through "dev" builds first and the case cache it
take under 3 seconds, for debug builds it takes closer to 30 seconds.

So, begin building the list of tests from dev and release builds, to
populate the case cache quickly and share the list among all
builds by default.

The caveat is that not all test files have the same list of tests
in release and debug modes (e.g. raft/many_test), for these files
introduce skip_case_cache options to suite.yaml.