test.py: harness for cluster management from test
Harness for managing cluster for Python tests.

Change cluster management from a list of ScyllaServer to Harness.
Provide tests a REST API with a unix socket for control.

Add simple topology tests with a simple schema change in test_topology.

Clusters can be marked recycled but are not yet recycled in test.py.
This will come in a later series.

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>