test.py: invoke exit artifacts ASAP
Don't wait until exit to invoke exit artifacts. These
artifacts stop running Scylla instances, so best to do it
as soon as the suite is done with, not at the end of the run.

This is especially important in CI when we run tests in
all suites in all build configurations.